add native asset proposal
