+++ scripts/js/setup.js
@@ -2,8 +2,21 @@ require('dotenv').config();
 
 const {ApiPromise, WsProvider, Keyring} = require('@polkadot/api');
 const {cryptoWaitReady} = require('@polkadot/util-crypto');
-const BN = require('bn.js');
+const {
+    setBalance,
+    setFeeHandler,
+    setMpcAddress,
+    registerDomain,
+    setFee,
+    getNativeAssetId,
+    createAsset,
+    setAssetMetadata,
+    mintAsset,
+    getUSDCAssetId,
+    queryBridgePauseStatus
+} = require("./util");
 
+const BN = require('bn.js');
 const bn1e12 = new BN(10).pow(new BN(12));
 
 const feeHandlerType = {
@@ -23,323 +36,7 @@ const supportedDestDomains = [
 ]
 
 const FeeReserveAccountAddress = "5ELLU7ibt5ZrNEYRwohtaRBDBa3TzcWwwPELBPSWWd2mbgv3";
-
-async function setBalance(api, who, value, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to set balance of ${who} to ${value}. (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.sudo
-            .sudo(api.tx.balances.setBalance(who, value, 0))
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function setFeeHandler(api, domainID, asset, feeHandlerType, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to set fee handler on domainID ${domainID}. (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.sudo
-            .sudo(api.tx.sygmaFeeHandlerRouter.setFeeHandler(domainID, asset, feeHandlerType))
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function setFee(api, domainID, asset, amount, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to set basic fee on domainID ${domainID}. (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.sudo
-            .sudo(api.tx.sygmaBasicFeeHandler.setFee(domainID, asset, amount))
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function setMpcAddress(api, mpcAddr, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to set MPC address. (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.sudo
-            .sudo(api.tx.sygmaBridge.setMpcAddress(mpcAddr))
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function queryBridgePauseStatus(api, domainID) {
-    let result = await api.query.sygmaBridge.isPaused(domainID);
-    return result.toJSON()
-}
-
-async function createAsset(api, id, admin, minBalance, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to create asset: (nonce: ${nonce}) ---`
-        );
-
-        const unsub = await api.tx.assets.create(id, admin, minBalance)
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function setAssetMetadata(api, id, name, symbol, decimals, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to register asset metadata: (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.assets.setMetadata(id, name, symbol, decimals)
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function mintAsset(api, id, recipient, amount, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to mint asset to ${recipient}: (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.assets.mint(id, recipient, amount)
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-async function registerDomain(api, domainID, chainID, finalization, sudo) {
-    return new Promise(async (resolve, reject) => {
-        const nonce = Number((await api.query.system.account(sudo.address)).nonce);
-
-        console.log(
-            `--- Submitting extrinsic to register domainID ${domainID} with chainID ${chainID}. (nonce: ${nonce}) ---`
-        );
-        const unsub = await api.tx.sudo
-            .sudo(api.tx.sygmaBridge.registerDomain(domainID, chainID))
-            .signAndSend(sudo, {nonce: nonce, era: 0}, (result) => {
-                console.log(`Current status is ${result.status}`);
-                if (result.status.isInBlock) {
-                    console.log(
-                        `Transaction included at blockHash ${result.status.asInBlock}`
-                    );
-                    if (finalization) {
-                        console.log('Waiting for finalization...');
-                    } else {
-                        unsub();
-                        resolve();
-                    }
-                } else if (result.status.isFinalized) {
-                    console.log(
-                        `Transaction finalized at blockHash ${result.status.asFinalized}`
-                    );
-                    unsub();
-                    resolve();
-                } else if (result.isError) {
-                    console.log(`Transaction Error`);
-                    reject(`Transaction Error`);
-                }
-            });
-    });
-}
-
-function getUSDCAssetId(api) {
-    return api.createType('XcmV1MultiassetAssetId', {
-        Concrete: api.createType('XcmV1MultiLocation', {
-            parents: 1,
-            interior: api.createType('Junctions', {
-                X3: [
-                    api.createType('XcmV1Junction', {
-                        Parachain: api.createType('Compact<U32>', 2004)
-                    }),
-                    api.createType('XcmV1Junction', {
-                        // 0x7379676d61 is general key of USDC defined in sygma substrate pallet runtime for testing
-                        // see UsdcLocation defination in runtime.rs
-                        GeneralKey: '0x7379676d61'
-                    }),
-                    api.createType('XcmV1Junction', {
-                        // 0x75736463 is general key of USDC defined in sygma substrate pallet runtime for testing
-                        // see UsdcLocation defination in runtime.rs
-                        GeneralKey: '0x75736463'
-                    }),
-                ]
-            })
-        })
-    })
-}
-
-function getNativeAssetId(api) {
-    return api.createType('XcmV1MultiassetAssetId', {
-        Concrete: api.createType('XcmV1MultiLocation', {
-            parents: 0,
-            interior: api.createType('Junctions', 'Here')
-        })
-    })
-}
+const TransferReserveAccount = "5EMepC39b7E2zfM9g6CkPp8KCAxGTh7D4w4T2tFjmjpd4tPw";
 
 async function main() {
     const sygmaPalletProvider = new WsProvider(process.env.PALLETWSENDPOINT || 'ws://127.0.0.1:9944');
@@ -385,8 +82,9 @@ async function main() {
         await setFee(api, domain.domainID, getUSDCAssetId(api), basicFeeAmount, true, sudo);
     }
 
-    // transfer some native asset to FeeReserveAccount as Existential Deposit(aka ED)
+    // transfer some native asset to FeeReserveAccount and TransferReserveAccount as Existential Deposit(aka ED)
     await setBalance(api, FeeReserveAccountAddress, bn1e12.mul(new BN(10000)), true, sudo); // set balance to 10000 native asset
+    await setBalance(api, TransferReserveAccount, bn1e12.mul(new BN(10000)), true, sudo); // set balance to 10000 native asset
 
     // bridge should be unpaused by the end of the setup
     for (const domain of supportedDestDomains) {
